% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxOT-package.R
\docType{package}
\name{approxOT}
\alias{approxOT}
\alias{approxOT-package}
\title{An R package to perform exact and approximate optimal transport.}
\description{
R and C++ functions to perform exact and approximate optimal transport. All C++ methods are linkable to other R packages via their header files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ericdunipace/approxOT}
  \item Report bugs at \url{https://github.com/ericdunipace/approxOT/issues}
}

}
\author{
Eric Dunipace
}
