% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hilbert_proj_}
\alias{hilbert_proj_}
\title{Returns orders along the Hilbert space-filling Curve}
\usage{
hilbert_proj_(A)
}
\arguments{
\item{A}{a matrix of data-values of class Eigen::MatrixXd}
}
\value{
An integer vector of orders
}
\description{
Returns orders along the Hilbert space-filling Curve
}
\keyword{internal}
