% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinkhorn.R
\name{round_transport_matrix}
\alias{round_transport_matrix}
\title{Round transportation matrix to feasible set}
\usage{
round_transport_matrix(transport_matrix, mass_x, mass_y)
}
\arguments{
\item{transport_matrix}{A transportation matrix returned by an approximate method}

\item{mass_x}{The distribution of the first margin}

\item{mass_y}{The distribution of the second margin}
}
\value{
Returns a transportation matrix projected to the feasible set.
}
\description{
Round transportation matrix to feasible set
}
\keyword{internal}
