% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinkhorn.R
\name{sinkhorn_distance}
\alias{sinkhorn_distance}
\title{Test sinkhorn distance}
\usage{
sinkhorn_distance(mass_x, mass_y, cost = NULL, p = 1, eps = 0.05, niter = 100)
}
\arguments{
\item{mass_x}{empiric measure of first sample}

\item{mass_y}{empiric measure of second sample}

\item{cost}{cost matrix}

\item{p}{power to raise the cost matrix by}

\item{eps}{epsilon of cost matrix}

\item{niter}{number of iterations}
}
\value{
a numeric value
}
\description{
Test sinkhorn distance
}
\keyword{internal}
