% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_plan.R
\name{sinkhorn_pot}
\alias{sinkhorn_pot}
\title{Return the dual potentials for the Sinkhorn distance}
\usage{
sinkhorn_pot(
  mass_x,
  mass_y,
  p = 2,
  cost = NULL,
  cost_a = NULL,
  cost_b = NULL,
  ...
)
}
\arguments{
\item{mass_x}{}

\item{mass_y}{}

\item{p}{}

\item{cost}{}

\item{cost_a}{}

\item{cost_b}{}

\item{...}{}
}
\value{

}
\description{
Return the dual potentials for the Sinkhorn distance
}
\keyword{internal}
