% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinkhorn.R
\name{sinkhorn_transport}
\alias{sinkhorn_transport}
\title{Test sinkhorn transportation plan}
\usage{
sinkhorn_transport(mass_x, mass_y, cost = NULL, eps = 0.05, niterations = 100)
}
\arguments{
\item{mass_x}{empiric measure of first sample}

\item{mass_y}{empiric measure of second sample}

\item{cost}{cost matrix}

\item{eps}{epsilon of cost matrix}

\item{niterations}{number of iterations}
}
\value{
transportation plan as list with slots "from","to", and "mass"
}
\description{
Test sinkhorn transportation plan
}
\keyword{internal}
